move helper functions into hf2 at cost of a few more deps
