+++ cargo-hf2/Cargo.toml
@@ -15,10 +15,7 @@ structopt = "0.3.2"
 colored = "2.0.0"
 hf2 = { version = "^0.2.1", path = "../hf2" }
 hidapi = "1.2.1"
-goblin = "0.2.3"
-hexdump = "0.1.0"
 cargo-project = "0.2.4"
 pretty_env_logger = "0.3.0"
 maplit = "1.0.2"
-crc-any = { version = "2.2.3", default-features = false }
 log = "0.4.6"
